testkit: Guard 'CPUSHA DC' with a 68040+ test.
The test with 'MOVEC ITT0' is safer than a bare F-Line instruction,
which results in a bus cycle on 020/030 CPUs and is subject to the
vagaries of accelerator glue logic.

In fact CPUSHA was reported to act unreliably on an Apollo 1230-Mk2
by Peter Mulholland. This workaround fixes his boot hang.

At the same time get rid of PMOVEs to clear 030 TT0,TT1 registers.
They are harmless to leave enabled.