fixes #24 - ATK - compile-error memcheck (memory.c) - 2nd attempt
replaced reference to "s" (char array) by using pointer to first
element. s is known to be in a safe memory area to write to, thus
the checks may start there to write the first patterns.
gcc 9.2.0 complains in using "s" in memory-pointer p with exceeding
array limits - using pointer to first element avoids the check.